Added PatientAge support
In previous iterations, if the user selected to anonymize by patient
birthday then it would remove the birthdate and not check to ensure that
the patient age was correct.

In this version, the birth YEAR is kept (ensuring that the age can be
computed properly) and the age is computed and the DICOM field
PatientsAge is created for patients where that field did not previously
exist.